Registrierung der Metadaten in die entsprechenden Unterklassen verschoben
