+++ src/Types/CustomType.php
@@ -1,6 +1,8 @@
 <?php
 namespace abrain\Einsatzverwaltung\Types;
 
+use abrain\Einsatzverwaltung\TaxonomyCustomFields;
+
 /**
  * Interface CustomType
  * @package abrain\Einsatzverwaltung\Types
@@ -16,4 +18,10 @@ interface CustomType
      * @return array
      */
     public function getRegistrationArgs();
+
+    /**
+     * @param TaxonomyCustomFields $taxonomyCustomFields
+     * @return void
+     */
+    public function registerCustomFields(TaxonomyCustomFields $taxonomyCustomFields);
 }
